pyproject.toml
[project]
name = "osu-pancakes"
version = "0.1.0"
requires-python = "3.13.12"
dependencies = [
  "blinker==1.9.0",
  "certifi==2026.2.25",
  "charset-normalizer==3.4.7",
  "click==8.3.2",
  "colorama==0.4.6",
  "dotenv==0.9.9",
  "Flask==3.1.3",
  "gunicorn==25.3.0",
  "idna==3.11",
  "itsdangerous==2.2.0",
  "Jinja2==3.1.6",
  "MarkupSafe==3.0.3",
  "osu.py==4.2.0",
  "packaging==26.0",
  "python-dotenv==1.2.2",
  "requests==2.33.1",
  "urllib3==2.6.3",
  "Werkzeug==3.1.8"
]

[project.scripts]
app = "main"